Change RAID command print priorities (jlaska).
We can change the priorities a bit more, but the issue is that
raid partitions must come before lvm stuff.  The failure encountered
here is when creating a kickstart to describe a "LVM on raid0" install.
Anaconda wasn't happy since the pv.01 (raid) partition was below the
volgroup that used it.