Add removedKeywords and removedAttrs lists on Commands and Data.
These new lists are for when a subclass removes support for a flag that was
supported in the superclass.  We can remove any trace of the older support
by ignoring the contents of removedKeywords in __call__ and __init__,
and by deleting everything in removedAttrs from the object itself.

Handling the __call__ and __init__ cases happens automatically at the
ultimate superclass, though the subclass must make sure to set up the two
lists before calling the superclass's __init__.  Removing the attributes
themselves requires an explicit call to deleteRemovedAttrs() after calling
the __init__ method.