Pass arguments to superclasses via *args and **kwargs, all the way up.
I've been meaning to do this for a very long time now.  Having explicit lists
of arguments on every single __init__ method in every subclass is hard to read
and hard to maintain.  Instead, pass everything via **kwargs and pull out what
we need at each step in the class chain.  It might be nice to have some
checking to make sure arguments that aren't recognized raise errors, but we
ignore this case everywhere else so why not here too?