Add upgrade --root-device (atodorov, #471232).
One outstanding problem with preupgrade initiated upgrades is that on machines
with multiple root partitions (because there are multiple operating systems
installed), anaconda will get confused about which contains the packages to
upgrade and which contains the filesystem to apply those packages to.  This
patch allows preupgrade to pass that information to anaconda.