Add support %include to the pykickstart data objects.
Previously, there was no way to represent the presence of a %include in
a kickstart file via the data objects.  pykickstart could process files
containing includes, but it couldn't do anything with them via the API.
This adds support for putting a %include before any command, script, or
packages section.  Includes cannot be added into the middle of a section
or at the very end of a file.