Move lineno= from KSOptionParser.__init__ to parse_args (#497149).
Setting it in __init__ causes all line numbers in error messages to be
reported as 0, since the option parsers are all instantiated when the
command handlers are created early-on.  Instead, we want line numbers to
reflect where we really are in processing, and that can only be done when
parse methods are called.