Support removing groups that were included by a glob (#554717).
pykickstart doesn't know anything about packages, and therefore does not
expand globs.  Group exclusion currently only works by taking into account
groups previously mentioned in the kickstart file.  This implementation
was primarily motivated by the spin kickstart files that %include.  So, a
kickstart snippet like the following:

%packages
*
-@foo
%end

will not exclude the group foo, because it was not previously mentioned.
This patch adds a new list that users of pykickstart can inspect to make
sure to remove the groups.