Don't overwrite the excluded group list after every %packages line (#577334).
Without this patch, you can only have one excluded group and it must come as
the very last line of the %packages section.  This is because after every
line of the %packages section, we're resetting the excluded group list back to
either whatever group you specified to exclude on this line, or nothing for
every other line.