If an option is deprecated, do not care if it takes a value (#602303).
The problem here is that if an option that ever did have a type that requires
a value (string, for instance) then when that option is later deprecated, it
will still require a value.  Then you'll get a parse error instead of the
deprecation warning, which doesn't make a lot of sense.