Make use of the "interactive" command an error.
Marking it as deprecated was a bit of a mistake.  If you had a generic
kickstart file with generic partitioning and expected interactive mode to
stop and prompt you, having it being deprecated would mean the installer
would just cruise on through anyway.  This could potentially damage any
installed system.