Allow for defining your own kickstart %sections.
This means consumers of pykickstart can create classes that define the behavior
of their own special sections, register them with the parser, and they will be
handled exactly the same as any of the builtin sections.

I'm hoping this will be useful for such things as the image generation tools.
I'm also considering this as a pykickstart-2.0 feature.  It certainly may
require moving objects around to break existing imports.