Add --nontp option and a way to specify NTP servers to the timezone command
We need a place to store NTP servers specified in the UI and it has to be
in ksdata, so let's allow users to specify them in kickstarts as well.
Since NTP is the default behaviour in the UI installation it should also
be the default for the kickstart installation. Thus there should be
--nontp option rather then --ntp option.