Raise an error if autopart is combined with partitioning commands (#886010)
Using autopart together with the part/partition, raid, logvol or
volgroup commands in a single kickstart file can lead to a very bad
and difficult to predict behaviour. It can cause hard to diagnose
installation errors or even make the installed system unbootable.
Therefore, better raise a parsing error and interrupt the installation
if autopart is combined with any of those commands.