Reset method seen attrs when switching method (#1004889)
The method initially gets set to cdrom, it can then be switched to
another method via UI or when falling back to closest mirror if a repo
isn't found. This would result in .seen to be set on multiple methods,
and since .cdrom.seen was checked first it would be used for the
attributes -- some of which don't exist.

This clears the .seen attr when switching methods or setting it to None.

It also fixes a problem where None was being returned for too many attrs
which caused problems with the internal operations of copy.copy used by
the software spoke to detect changes from the previous method settings.
Only return None if the name is "method" otherwise pass it through to
the superclass' __getattr__