method getattr should default to handler.url
When setting method attributes it defaults to using the url handler when
method is None, but there was no way to retrieve them once set. This
re-factors the __getattr__ handler so that internal methods go to the
right place and all others go to the first seen handler, or to
handler.url if method is None (used for closest mirror).