Set bootloader location constructor default value to "none" (#916529)
Related: rhbz#916529

Bootloader location default set in constructor was previously "", which
is not one of the allowed choices; client code
is not likely to expect or check for this value.

There is a possiblity that some client code somewhere uses
bootloader.location=="" as a proxy for not bootloader.seen, in which case
it will break after this change.

Signed-off-by: mulhern <amulhern@redhat.com>