Move commandMap and dataMap setting into the individual handler classes.
This is a step towards consolidating what must be edited in order to make a new
version of kickstart syntax.  Now you keep the commandMap and dataMap together
in the WhateverHandler class and it gets built into the master dicts for you.

There should be no changes visible to users of pykickstart, though.  The master
commandMap and dataMap should still be the same.