Make print statements Python 3 compatible
The print statement has been dropped in Python 3.
Fortunately, if the print statement is written as if it were a
function (print("foo") vs print "foo") it still works in the same way, is
Python 3 compatible and the few special cases can be handled by importing the
print_function from the __future__. So there is really no reason to
not just use the print "function" everywhere and make Pykickstart a bit
more Python 3 friendly.