Move some statically detectable kickstart errors out of anaconda (#1117908)
Related: rhbz#1117908

These errors have everything to do with the options in the file and nothing
to do with the situation in which they are supposed to be executed, so they
should be checked when the file is read.

Suggest --noformat option as well, since the effect of --noformat option is
to set preexist to True.

Signed-off-by: mulhern <amulhern@redhat.com>