Keeping order of contents in the %packages section with OrderedSet (under MIT license).
Pykickstart's parser was dependent on the fact that sets in Python 2 usually
kept the order of the elements, which made the %packages section work as
expected. In Python 3, the order is randomized, therefore using an
implementation of ordered sets is necessary.

The OrderedSet recipe by Raymond Hettiger is used for this, taken from
http://code.activestate.com/recipes/576694-orderedset/, released
under the MIT license.