Replaced URLGrabber with requests (rhbz#1141245)
Because urlgrabber is being dropped by many significant software projects in
Fedora, Pykickstart can't keep depending on it once it is converted to Python 3
(rhbz#985310).

This commit replaces the dep with requests with SSL verification turned on by
default. This could pose trouble in Python 2 if proper underlying packages
aren't installed [1]. In case this becomes a problem, the SSL_VERIFY constant
in load.py may be switched to False.

[1] https://github.com/kennethreitz/requests/issues/749#issuecomment-19187417