Make pocketlint optional build dependency
As of Fedora23 all core packages runs with python3,
therefore we would like to keep build root for core
packages as small as possible to make the rebuilds
easier of Python3 easier. Please note that pocketlint
doesnt run currently during the build process so its
listing as BuildRequires is redundant.