Add some initial static typing information to pykickstart.
Add static typing information to the public methods on pykickstart.base
and pykickstart.version, since that seems like as good a place as any to
start, and any of the collection types where mypy needs type
annotations. Run mypy during 'make test' and add enough 'type: ignore'
comments to make it work.

All of the type information is specified as comments instead of using
the Python 3 type annotation syntax because pykickstart needs to be able
run as Python 2. The type comments are kind of gross in that they need
to be crammed all on to one line, but they seem a lot less likely to
make everything explode than the mypy codec provided as an alternative.

Fixed one actual type problem: pykickstart.base.BaseHandler initializes
currentLine as an int, but everywhere else this value is a string.
Initialize it to "" instead.