It's self.sshkey, not self.key.
This was passing tests because when the --sshkey argument is given,
we'll get op.sshkey set with the value.  That in turn then gets set to
self, so there's a self.sshkey.  self.key is still there and set to its
default value, but we never check it anywhere.

Meanwhile, we're never testing the setting of attributes by passing
things in to __init__.  That's the only way we'd ever see that this was
broken.