Add a syntax highlighting file for vim.
We've actually had this for a while, but it's time the rest of the world
can make use of it.

I'm just installing this in the docs directory for now because I don't
want to deal with packaging.  You can make use of it by either copying
it into /usr/share/vim/vimfiles/syntax/ globally, or ~/.vim/syntax on a
per-user basis.