Make sure the script test references parser.
What I was seeing was that the test failed because the script strings
didn't have a %end on them.  That means the version was being set to
something pre-F8.  It only happened if any of the pre* or post* tests
were run first - if those were moved out of the way, the script test
would pass.  These tests were therefore the ones causing the version to
be set incorrectly.

What's happening is those tests set the version to something old, but
nothing sets the version back.  That's because Script references a
global ver variable which is set when a parser is created.  The Script
test doesn't create a parser so ver was staying set to whatever old
value pre* and post* were setting.

(cherry picked from commit 5c33825407035f33bfd8ae257ec093c10937b81c)