Get rid of the ver global variable.
This was only used for knowing whether to print out %end for scripts and
packages, which means now it's currently only used for when you are
processing very old kickstart files.  I've moved it in to being a class
attribute on Script and Packages.

While I was at it, add some more test cases around this stuff and fix a
bug where scripts were getting too many newlines added after them.