Add a test case for various ways of setting attributes.
pykickstart command and data objects can have their attributes set in a
variety of ways.  The most obvious is by parsing a kickstart file.  But
you can also pass them as kwargs when you __init__ an object and you can
also pass them as kwargs if you __call__ a previously instantiated
object.  Until now, there's been no testing of those other two methods.

This adds that capability.  It's not perfect - it only knows about
attributes that can be set via some argument and the parse method - but
it's pretty good.