Disable the attrs test for python2.
It doesn't work, and I don't really care to figure out why because
python2 is on the way out on this branch anyway.  However, I also
couldn't make "-I attrs.py" work so I'm doing this the hard way and
using make to figure it out.  The result is the same.