Lots of documentation updates.
Remove all the references to when something was introduced.  A lot of these were from back in the Fedora 16 days.  This material belongs in release notes or the installation guide.  It doesn't belong on the master branch for all time.

Second, remove almost all of the bolded text.  There was entirely too much and I don't think most people read documentation especially closely anyway, and are pretty much trained to skim over large blocks of bold.

Third, remove the monitor command reference.  That command was deprecated in F10.