Support timezone command usage without timezone specification
Make it possible to use the timezone command without a timezone
specification.

This way the timezone command can be used with the --utc flag
to set hardware clock time format while still requiring the user
to select a timezone manually in the Anaconda UI.

Also throw an exception if only "timezone" is used (eq. no options &
arguments) as that really does not make sense.

Related: rhbz#1312135