New version.
Note: the past couple commits are out of order, due to me not pushing
the previous "New version" commit when I did it.  The following:

* a7b4e82 - New version.
* 72466e8 - Support timezone command usage without timezone specification
* ef6dfa0 - Add --no-activate option to network command (#1277975)
* 73c9df1 - (tag: r1.99.66.7-1) New version.

Ought to have been in this order:

* 72466e8 - Support timezone command usage without timezone specification
* a7b4e82 - New version.
* ef6dfa0 - Add --no-activate option to network command (#1277975)
* 73c9df1 - New version.

That matches up the order the commits were actually done in, as well as
what fixes are in what release.