Test for older versions in new Fedora releases. Closes #28
This test loops over the contents of commandMap and dataMap and
tries to find handlers (e.g. Fedora versions) for which the listed
commands and data attributes are not the latest possible version
defined in the corresponding command.py file.

For example if the latest release is Fedora24 and we're testing the
handler for Fedora23's "logvol" then we expect the mapping to be
F23_LogVol or lover, even if F24_LogVol is present. However if the
mapping is F20_LogVol then this is a bug b/c F23_LogVol provides
newer implementation which isn't used.

The test doesn't report errors for RHEL vs Fedora differences and
vice versa.