Delete str != "" comparisons to remove 8*110 possible mutations
During mutation testing `if str != ""` becomes
`if str > ""` or `if str not in ""` which are mutations we can't kill!

Every time str is not an empty string the 3 variants of the condition
are evaluated to True and the existing test case passes. To kill the
mutants we expect the test to fail but that never happens b/c in this
case all 3 mutations are logically equivalent!

This change also greatly reduces the number of possible mutations to
about 4,4K instead of 5,5K.