Don't skip DeprecatedCommands when testing handler mappings
as is evident there were a few cases where a command has been
deprecated in an earlier release but later releases used a
non-deprecated version of the command.

We also want to make sure the devel handler always has the latest
command versions in order to make it easier for kickstart docs
generation.