Split upgrade and install commands and update handlers after F20
In Fedora 20 the upgrade part of the install|upgrade command has
been deprecated. However this interferes with
HandlerMappingTestCase b/c it doesn't skip deprecated commands
anymore and tells us the "install" : commands.upgrade.F11_Upgrade
mapping is wrong.

By moving the install implementation into it's own class we make
the test happy and provide clear path for future implementations
of the install command if need be.

We also make it easier for the automatic kickstart documentation
generator b/c it relies on having all the latest command versions
in the devel handler.