Automatically build kickstart command & sections documentation
also add testing docs to index.rst and update formatting to avoid
Sphinx warnings. We make sure to document commands which may no
longer be in use by the latest handlers.

Note that we build command docs separately for RHEL and Fedora!

Also note that we don't use .. versionadded:: for descriptions of
packages and sections since it always points to DEVEL.