Fix F25 timezone should subclass F23 instead of F18
The F23_Timezone is required because of ntp pools. NTP pools are now
specified with repeating of the same ntp server 4 times.

The problem here is that in F18_Timezone there is set() for storing
ntp_servers which can't store multiple times the same server. The
variable was changed to list() in F23_Timezone.