Retroactively fix checks for reqpart and autopart
commit 97a84e5, rhbz#1164660 adds conflict check between
reqpart and autopart commands but does so not very nicely.

The check is added to the RHEL6_AutoPart and F20_AutoPart
classes which already deal with similar conflicting conditions.
At the same time the F23_ReqPart command was added to f23.py
handler.

This commit removes the 'reqpart' check from the RHEL6_AutoPart
class because this is never used. It adds an F23_AutoPart class
for this new check and updates f23.py and later handlers. It also
updates the RHEL7_AutoPart class and adds more tests.