Warn about using removed keywords in kickstart commands
In KickstartCommand.__init__() the code

    for arg in (kw for kw in self.removedKeywords if kw in kwargs):
        kwargs.pop(arg)

doesn't do what it is intended to do. kwargs is passed by value,
not by reference and although we remove the unnecessary parameters
this doesn't propagate to the inherited class. This can easily be
demonstrated with the following test:

class KickstartCommandWithRemovals(KickstartCommand):
    removedKeywords = KickstartCommand.removedKeywords + ["connect"]
    removedAttrs = KickstartCommand.removedAttrs + ["connect"]

    def __init__(self, writePriority=0, *args, **kwargs):
        KickstartCommand.__init__(self, writePriority, *args, **kwargs)
        kwargs.pop('connect') # should raise KeyError

class RemovedKeywords_TestCase(unittest.TestCase):
    def runTest(self):
        with self.assertRaises(KeyError):
            KickstartCommandWithRemovals(connect='test')

The above test doesn't raise a KeyError because in
KickstartCommandWithRemovals.__init__() kwargs still contains the
'connect' key/value pair.

Instead the classes which remove attributes need to call
`self.deleteRemovedAttrs()` inside the constructor to deal with the
removals. This commit adds a test case for this as well.