Refactor in dmraid.py
before this change the line

	dm.name = dm.name.split('/')[-1]

didn't really do anything because dm.name was empty string. This is
because the dm object has just been constructed and no values
have been assigned to it! As far as I understand the intention is
to be able to accept names like "/dev/something" and strip off
everything before the last slash which this change actually
acomplishes.

Another alternative is to apply the following patch instead:

-	dm.name = dm.name.split('/')[-1]
+	dm.name = ns.name.split('/')[-1]