Run the docs makefile during RTD build
Run the docs makefile to generate intermediate documentation files
needed for a successful Read the Docs build.

This is achieved by running the make file from the conf.py script
which is executed during the RTD build. The second step needed is
is to visit advanced settings for the corresponding RTD project and:
- enable the "Install your project inside a virtualenv using setup.py install" option
- set the "Requirements file:" field to "docs/rtd-requirements.txt"

Also remove the pre-generated rst files, because these automatically generated
during the RTD build and no longer need to be stored in the repo.