Disable running "make coverage" or "make check" with python2.
We are using unittest.mock in a very large number of places now, and
that doesn't exist in the python2 version of unittest.  For a while I
could just get away with disabling the tests that imported this module,
but it now appears to be a very large portion of them.

I've been looking to phase out python2 support on this branch anyway, so
this is no big loss to me.