Pass comments=True to shlex.split calls in the test functions.
This mirrors how it's being called in the real parser.  Not doing this
was hiding a bug where parsing passwords with a # in them apparently
worked in testing, but failed in real use.  The testing only apparently
worked because the expectedStr also gets parsed, and drops comments, so
both sides were dropping comments in the same way.

In other words, I was passing "comment#password" into both sides, it was
getting turned into "comment" in each, and those matched.  I need to
also be checking that rootpw.password is "comment#password".

So with this commit, I get the expected failure.

Related: rhbz#1301267