Fix handling # in passwords.
See the previous commit for why this was still broken.  I've also added
a test case to make sure it's fixed this time and stays that way.

Note that this commit introduces new behavior:  If your password
contains a #, the password must be surrounded by quotes or it'll be seen
as a comment and stripped out.  If you generate a password with a # in
it, pykickstart will quote it for you.  This seems like a reasonable
thing to do, as both handling and ignoring end-of-line comments in this
one case is pretty difficult.

Related: rhbz#1301267