Add kickstart warnings
Pykickstart should use its own warning categories for warnings related
to parsing a kickstart file, so it will be possible to distinguish
them from warnings related to API changes.

For example, it wasn't possible to tell if the DeprecationWarning
is related to a deprecated kickstart command or a deprecated
pykickstart method.

All warnings related to parsing should use the KickstartParseWarning
category. All warnings related to deprecated commands and options
should use the KickstartDeprecationWarning category.