Apply a different filter for warnings from pykickstart (#1408667).
Using "pykickstart" as the filter means that PendingDeprecationWarnings
coming from inside pykickstart still get printed.  I don't want people
to see those warnings that are a result of something in pykickstart
continuing to use a deprecated function.  I only want people to see the
warnings from their own code.

The new filter of "." gets rid of these.  I am hoping it does not also
get rid of the ones the user is responsible for.