Handle error message formatting in KickstartError
In anaconda, we need to be able to access the original error
message and the line number where the error occured. It wasn't
possible when KickstartError received only the formatted message.

Now, KickstartError handles the error message formatting and it
keeps the original message and the line number. The function
formatErrorMsg is deprecated.

The recommended way of creating an exception:

    KickstartError(message, lineno=lineno)

The deprecated way is still supported:

    KickstartError(formatErrorMsg(message, lineno=lineno))