Remove the spec file from the source repo.
They keep changing the spec file in Fedora package source control, which
means I have to remember to sync that up with what's in pykickstart git
or I lose their changes.  I'm tired of that, and having the spec file in
two places is just dumb anyway.  This removes it from upstream git and
leaves it in the distribution's packaging source control where it
belongs.