Fix 'make rpmlog' target.
This one was irritating me.  Worked fine if you ran it before making a
release, but once a new tag is in place it returns nothing.  Modified it
to return an rpm spec file changelog block for the commits from the
previous tag to the current tag.

(cherry picked from commit 8cdf59f1a09306735d8d9fff3470da1cc4712958)