Do not throw exception on deprecated ignoredisk --interactive
The option was deprecated by
commit 67bde6b1bfe0a20d17a95620d56df96dfc805efej
but we don't want to abort the installation for existing kickstarts using the
option so loose the guard in parse method and add a deprecation warning.

If there is the --interactive option only, just issue the warning, do not raise
parse error. (These kickstarts used to be accepted.)

If there is also other option present in the command, modify the message of
raised exception with information that the --interactive option is no longer
supported. (These kickstarts used to raise the exception but the message should
be changed.)

Resolves: rhbz#1637872