Fix warnings for deprecated options of kickstart commands
If we try to parse a deprecated option of a kickstart command, the
option parser should issue a warning. This functionality was broken
because of kickstart sections.

Kickstart commands are defined by classes specific for the given
version. In their case, the option parser don't have to check the
defined options against the current version of the kickstart syntax.
If the option is valid, then it is defined. If the option is removed,
then it is not defined. If the option is deprecated, then it is marked
as deprecated.

Kickstart sections are defined by classes common for all versions.
In their case, the option parser was expected to check the defined
options against the current version. However, that broke the detection
of deprecated options for kickstart commands.

Since the support for kickstart sections is useful only for the
section %packages, let's drop it, fix the support for kickstart
commands, remove the workarounds for sections and modify the sections.