Make %packages --default attribute sane
The %packages --default will now force default environment. User can't
change the environment even when he thinks he did. As a bonus
Anaconda will use the user-set environment everywhere but not for
installation.

All groups and packages written in the %packages section are installed
correctly but they are not propagated to the output kickstart file.

In this commit I'm changing the output section generator to print all
the groups and packages to the output kickstart but avoiding
environment, which will be forced and can't be changed. Anaconda may
change this behavior in the future but in that case the --default is
not used and will be disabled.

The other thing is that %packages --weakdeps (or any other attribute
except --default) without packages content has meaning but it won't be
printed in the output kickstart file.

Also change documentation to mention that only environment is set
nothing else.