Restrict the branches to test with Python 3.
This is where things are getting complicated with Travis-CI.
pykickstart maintains multiple concurrent stable branches.  master is
where development happens and it is currently using Python 3.  There is
a rhel7-branch for RHEL-7 and it uses Python 2.  We need automated tests
on all current branches, so this is an effort to enable that for
Travis-CI.