Give output in str(F25_Timezone) even without timezone spec
The previous implementation of F25_Timezone inherited the __str__() and
_getArgsAsStr() from F18_Timezone verbatim. F18 enabled output only if
timezone was specified. F25 made the timezone optional, but did not override
or otherwise modify the F18 output mechanism. This caused a discrepancy in F25
between what was allowed as input (no timezone spec) and what was treated
as situation where output makes sense (timezone required).

This fixes the problem by changing the F25 methods.

F25 timezone tests are also adjusted accordingly.