Fix travis and coverage runs
coveralls requires a specific range of coverage versions, if you also
install coverage it pulls in the wrong version. So just tell it to
install coveralls and let it pull in the compatible version.

after_success in travis already has the required packages installed, no
need to run pip again.