tests: Fix the empty description, help, etc. test
It was stopping on the first error in the class, which masked other
errors. It was also not showing which argument was triggering the error.

It now prints the class and all the arguments that fail so that it is
clear what needs to be fixed.

This test has been re-enabled.