Switch to using Weblate for translations
Fedora is moving to a new service which will commit new translations
directly into https://github.com/pykickstart/weblate

This commit switches po-pull to check out this repo and copy the .po and
pykickstart.pot file into ./po/ before making a new release.

It also adds a po-push target which will need to be run when strings
change. It will generate a new pykickstart.pot and push it to the
pykickstart weblate repo if it has changed. The pykickstart.pot file
will no longer be carried in the ./po/ directory.

The weblate repo branch should match the current pykickstart branch, but
it can be overridden by setting WEBLATE_BRANCH=foo when running make
po-pull or po-push.

Resolves: rhbz#1790574