docs: Maintain programmers-guide in the repo
This was being pulled from the Fedora wiki, but has remained unchanged
since it was first posted. pykickstart docs are now separate from Fedora
so it is easier to maintain this here in the repo.

This also removes the need for network access while building, which is
usually not available.