Add the timesource command
With the effort to add support for NTS (secure NTP) to Fedora
we have found out the current way of specifying NTP servers
via the --ntpservers option of the timezone command is no longer
sufficient. Not only can't it be easily and cleanly extended to also
support NTS servers, we are actually missing some options that
are available in the GUI, such as NTP pool support.

So add a new command called "timesource" that makes it possible to add
multiple NTP servers/pools and to specify if the given server/pool
supports NTS.

This way we can add support for NTS in a robust manner & make it easy
to add support for other time synchronization features as needed.

The timesource command can be used multiple times, one NTP/NTS server/pool
per line.

Options at a glance:

--ntp-server

Address of a NTP server.

--ntp-pool

Address of a NTP pool.

NOTE: It is not possible to specify both --ntp-server and --ntp-pool at the same time,
as the result would be ambiguous.

--nts

Specifies that the given server or pool supports the NTS protocol.

--ntp-disable

Specifies that all ntp based time synchronization should be disabled on the system.
If timesource is specified multiple times and at least one invocation uses the
--ntp-disable option, ntp based time synchronization will be disabled.