parser: Remove OrderedSet
We would like to reduce the dependency on other libraries, and
OrderedSet doesn't appear to really be necessary if the packageList and
excludedList is always sorted. pykickstart already sorts the package
list when returning the contents as a string so the only possible thing
this could effect is users who access the lists directly (eg. anaconda,
lorax-composer) and in those cases dnf should not care about the package
order.