test: Update Dockerfile to use CentOS 8
Fedora 29 is EOL and we now have CentOS available to use for running the
tests. This primarily effects the Travis tests run by the github pull
requests.

Use packages for as much of the test requirements as possible. This
means adding the EPEL8 repository and trimming the pip3 install to only
2 packages, pocketlint and Sphinx.

Resolves: rhbz#xxxxxx