Add RemovedCommand for removing commands and documenting them
Previously removing a command was done by removing it from the handler
list. But this left the command in the documentation with no note about
when it was actually removed.

With this class you can now create a new command subclass of
RemovedCommand and update the description with the version that it is
being removed in.

Any use of the command will raise an error, and ksverdiff has been
adjusted to check for commands removed from the handler and for any
commands that are subclasses of RemovedCommand.