Set default version to RHEL9
The default version on RHEL 9 should be RHEL9. But you cannot just set
DEVEL=RHEL9 because that will cause other problems and confusion.

So, this sets the default for files and makeVersion to RHEL9 which will
mean that the correct options are supported without passing a specific
version to the functions.

These can still be overridden by passing a different version, or setting
the version in the file comments.

Related: rhbz#2058335