Write commands in alphabetical order
When a kickstart is output the commands are grouped by write priority,
then alphabetical within the group. This isn't how it was actually
working -- it was using the full class name with release prefixes (eg.
F37_) so a newer command would sort before older commands, as seen with
the recent rootpw change.

This changes the insert to compare based on the command name portion of
the class name, without the prefix. This will change the order of some
things -- but in the future the output will be in stable alphabetical
order.

Resolves: rhbz#2096871