Add new ostreecontainer command (#2125655)
It is another way to set source for ostree.

The difference:
ostreesetup - work for OSTree repositories
ostreecontainer - work for native container images

See https://fedoraproject.org/wiki/Changes/OstreeNativeContainer for
more details.

Changes:
The ostree tool renamed `osname` to `stateroot` so when we creating a
new command let's make the name correct.

The default value is recommended on BZ and here
https://github.com/ostreedev/ostree/issues/2794

Related: rhbz#2125655

Rename osname to stateroot and set default value (#2125655)

The ostree tool renamed `osname` to `stateroot` so when we creating a
new command let's make the name correct.

The default value is recommended on BZ and here
https://github.com/ostreedev/ostree/issues/2794

Related: rhbz#2125655