Do not request sign for rc-release Makefile target
Without this the rc-release Makefilet target is doomed to fail. The
point is that it requires creation of the tarball but won't give you an
opportunity to do the signing yourself before starting to consume
created tarball.

To fix this make the signing optional in spec file and build without the
sign verification.