Add conflicting commands support
Previously a command would have to implement checking the .seen
attribute on commands that it conflicts with, and manually update the
command description with the commands that it could not be used with.

With this change you can implement command conflict support by adding a
list to the class:

  conflictingCommands = ["raid", "reqpart"]

Pass this list to the command's KSOptionParser as ``conflicts``:

  KSOptionParser(..., conflicts=self.conflictingCommands)

And then in the first version that the conflict occurs, call the helper
function with a translated error message from the ``parse`` function:

  self._checkConflictingCommands(_("The %s and autopart commands can't be used at the same time"))

the command's description will have a list of the conflicting commands
appended to it automatically.

(cherry picked from commit 5f6c0454261d9ab65b318d8b86c509ae91336c54)
Related: RHEL-17662