Do not remove "md" prefix from RAID names
This used to be done when MD names were based on minor to remove
the prefix from names such as "md0". This is no longer true and
parsing the name in this way was removed in Fedora 19 (see
d55f1adba0bdf73fc41d88e0632dfb0dbf418e8e). Removing the prefix
from the name now causes issue when users use names like for
example "md_xyz" and then try to use the name to refer to the new
array by name in other commands like volgroup.

Resolves: RHEL-80086